Logo change, box behind links, shapes
Created a new logo in Illustrator which I placed on top of the 3DS image using a new div element, used that same element to then organize the navigation bar + the box behind it + paragraph under logo so that when the page is resized, those elements move as well. I added triangle shapes which are invisible until you hover over the link that corresponds to its triangle -- this was also in the .container class. I had to make multiple classes to do this (not sure if I did it the most efficient way).