Changed logo + bkg, added animation
Instead of adding a new file for the background, I included code that made it gradient (which matches with the top screen of the 3DS for aesthetic cohesiveness). I changed the logo so that it matches the bottom screen of the 3DS (included moon instead of rose bc more relevant). Added keyframes so the text under my logo blinks so it mimics a 3DS more. Resized the container for the navigation bar so the triangles show up when you hover over the text (before, the container was really big and the triangles appeared when the cursor wasn't over the text).